+++ src/common/hooks/useVendorCodeGeneratorSettings/index.js
@@ -1,0 +1,1 @@
+export { useVendorCodeGeneratorSettings } from './useVendorCodeGeneratorSettings';